+++ KCore/KCore/Nuga/include/clipper.hxx
@@ -988,13 +988,13 @@ namespace NUGA
         }
       }
 
-      int nbits = crd_res.size();
+      size_t nbits = crd_res.size();
       bits.reserve(nbits);
       std::move(ALL(crd_res), std::back_inserter(bits));
       // pass the history
-      for (int k = 0; k < nbits; ++k)
+      for (size_t k = 0; k < nbits; ++k)
       {
-        assert(poids_res[k].size() == bits[k].m_crd.cols());
+        assert(poids_res[k].size() == (size_t)bits[k].m_crd.cols());
         bits[k].m_poids = poids_res[k];
       }
